+++ src/hooks/useRafState.js
@@ -1,0 +1,16 @@
+import { useCallback, useRef, useState } from 'react';
+
+const useRafState = (initialState) => {
+  const frame = useRef(0);
+  const [state, setState] = useState(initialState);
+  const setRafState = useCallback((value) => {
+    cancelAnimationFrame(frame.current);
+    frame.current = requestAnimationFrame(() => {
+      setState(value);
+    });
+  }, []);
+
+  return [state, setRafState];
+};
+
+export default useRafState;
